Add files via upload
Assignment Details: 
Create a page that has 5 textboxes to hold 5 test scores.  Have a button that when clicked puts all of the scores into an array, then loops through the array and finds the average of all but test #3 (you must use the continue statement to skip test 3).  You may use document.write to inform the user what the average is and that test #3 is NOT part of the computation.